src/main/java/players/Player.java
package players;

import interfaces.takeDamagable;

public abstract class Player implements takeDamagable{

    private String name;
    protected int healthPoints;

    public Player(String name, int healthPoints) {
        this.name = name;
        this.healthPoints = healthPoints;
    }

    public String getName() {
        return name;
    }


    public int getHealthPoints() {
        return healthPoints;
    }

    public void takeDamage(int damage) {
        this.healthPoints -= damage;
    }


}
